refactor : Event 파일들 Discount , Present 로 구분
더 보기 쉽게 하기 위해 폴더별 이벤트 분리
