+++ src/main/java/christmas/event/discount/WeekdayDiscountEvent.java
@@ -1,6 +1,4 @@
-package christmas.event;
-
-import static christmas.constant.EventConstant.WEEKEND_DISCOUNT_PRICE;
+package christmas.event.discount;
 
 import christmas.domain.Date;
 import christmas.domain.DayOfWeek;
@@ -12,14 +10,16 @@ import christmas.lib.event.DiscountEvent;
 
 import java.util.List;
 
-public class WeekendDiscountEvent extends DiscountEvent<Bill> {
-    private final Integer DISCOUNT_PRICE = WEEKEND_DISCOUNT_PRICE;
-    private final String EVENT_NAME = "주말 할인";
+import static christmas.constant.EventConstant.WEEKDAY_DISCOUNT_PRICE;
+
+public class WeekdayDiscountEvent extends DiscountEvent<Bill> {
+    private final Integer DISCOUNT_PRICE = WEEKDAY_DISCOUNT_PRICE;
+    private final String EVENT_NAME = "평일 할인";
 
     @Override
     public boolean checkCondition(Date date) {
         DayOfWeek dayOfWeek = date.dayOfWeek();
-        if (dayOfWeek.isWeekend()) {
+        if (dayOfWeek.isWeekday()) {
             return true;
         }
         return false;
@@ -27,7 +27,7 @@ public class WeekendDiscountEvent extends DiscountEvent<Bill> {
 
     @Override
     public DiscountEventReward provideReward(Bill bill) {
-        List<OrderInfo> orderInfos = bill.orderDetail().get(Category.MAIN_DISH);
+        List<OrderInfo> orderInfos = bill.orderDetail().get(Category.DESERT);
         Integer count = countOrderMenu(orderInfos);
         return new DiscountEventReward(EVENT_NAME, count * DISCOUNT_PRICE);
     }
