+++ src/main/java/christmas/service/EventService.java
@@ -5,7 +5,11 @@ import static christmas.constant.EventConstant.EVENT_THRESHOLD_PRICE;
 import christmas.domain.Date;
 import christmas.domain.event.Reward;
 import christmas.domain.order.Bill;
-import christmas.event.*;
+import christmas.event.discount.ChristmasDiscountEvent;
+import christmas.event.discount.SpecialDiscountEvent;
+import christmas.event.discount.WeekdayDiscountEvent;
+import christmas.event.discount.WeekendDiscountEvent;
+import christmas.event.present.ChampagnePresentEvent;
 import christmas.lib.event.DiscountEvent;
 import christmas.lib.event.EventReward;
 import christmas.lib.event.PresentEvent;
