+++ src/test/java/christmas/event/ChampagnePresentEventTest.java
@@ -1,5 +1,6 @@
 package christmas.event;
 
+import christmas.event.present.ChampagnePresentEvent;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
