+++ src/test/java/christmas/event/ChristmasDiscountEventTest.java
@@ -4,6 +4,7 @@ import static christmas.constant.EventConstant.D_DAY_START_PRICE;
 import static christmas.constant.EventConstant.CHRISTMAS_DAY;
 
 import christmas.domain.Date;
+import christmas.event.discount.ChristmasDiscountEvent;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
