+++ src/test/java/christmas/event/WeekdayDiscountEventTest.java
@@ -4,6 +4,7 @@ import christmas.constant.EventConstant;
 import christmas.domain.Date;
 import christmas.domain.order.Bill;
 import christmas.domain.order.RequestOrder;
+import christmas.event.discount.WeekdayDiscountEvent;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
