+++ src/test/java/christmas/event/WeekendDiscountEventTest.java
@@ -4,8 +4,8 @@ import christmas.constant.EventConstant;
 import christmas.domain.Date;
 import christmas.domain.order.Bill;
 import christmas.domain.order.RequestOrder;
+import christmas.event.discount.WeekendDiscountEvent;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
